+++ src/libutil/unix/signals.cc
