+++ src/libutil/unix/signals.hh
